Feature/get all recipies

+++ backend/src/main/java/org/example/backend/model/Ingredient.java
@@ -1,4 +1,4 @@
 package org.example.backend.model;
 
-public record Ingredient(String name, Integer quantity, String unit, String additionalInfo) {
+public record Ingredient(String name, double quantity, String unit, String additionalInfo) {
 }
